Conflicts:
	server/amee-calculation/src/main/java/com/amee/calculation/service/CalculationConsumer.java
	server/amee-calculation/src/main/java/com/amee/calculation/service/CalculationService.java
	server/amee-calculation/src/main/java/com/amee/calculation/service/ProRataProfileService.java
	server/amee-core/src/main/java/com/amee/core/ValueType.java
	server/amee-core/src/test/java/com/amee/core/ValueTypeTest.java
	server/amee-domain/src/main/java/com/amee/domain/AMEEEntity.java
	server/amee-domain/src/main/java/com/amee/domain/data/ItemDefinition.java
	server/amee-domain/src/main/java/com/amee/domain/data/ItemValue.java
	server/amee-domain/src/main/java/com/amee/domain/data/ItemValueDefinition.java
	server/amee-domain/src/main/java/com/amee/domain/path/PathItem.java
	server/amee-domain/src/main/java/com/amee/domain/profile/ProfileItem.java
	server/amee-domain/src/main/java/com/amee/domain/profile/builder/v1/ProfileItemBuilder.java
	server/amee-domain/src/main/java/com/amee/domain/profile/builder/v2/ProfileItemBuilder.java
	server/amee-domain/src/main/java/com/amee/domain/sheet/Cell.java
	server/amee-domain/src/main/java/com/amee/domain/sheet/Choice.java
	server/amee-domain/src/main/java/com/amee/domain/sheet/Choices.java
	server/amee-domain/src/main/java/com/amee/domain/sheet/Sheet.java
	server/amee-restlet/src/main/java/com/amee/restlet/data/DataItemResource.java
	server/amee-restlet/src/main/java/com/amee/restlet/environment/AlgorithmResource.java
	server/amee-restlet/src/main/java/com/amee/restlet/profile/ProfileCategoryResource.java
	server/amee-restlet/src/main/java/com/amee/restlet/profile/ProfileItemResource.java
	server/amee-restlet/src/main/java/com/amee/restlet/profile/acceptor/ProfileCategoryFormAcceptor.java
	server/amee-restlet/src/main/java/com/amee/restlet/profile/builder/v1/ProfileCategoryResourceBuilder.java
	server/amee-restlet/src/main/java/com/amee/restlet/profile/builder/v2/ProfileCategoryResourceBuilder.java
	server/amee-restlet/src/main/java/com/amee/restlet/profile/builder/v2/ProfileItemResourceBuilder.java
	server/amee-service/src/main/java/com/amee/service/profile/ProfileBrowser.java
	server/pom.xml

+++ server/amee-restlet/src/main/java/com/amee/restlet/environment/AlgorithmResource.java
@@ -19,13 +19,13 @@
  */
 package com.amee.restlet.environment;
 
-import com.amee.platform.science.CO2Amount;
 import com.amee.domain.AMEEEntity;
 import com.amee.domain.algorithm.Algorithm;
 import com.amee.domain.algorithm.AlgorithmContext;
 import com.amee.domain.profile.ProfileItem;
 import com.amee.domain.sheet.Choice;
 import com.amee.platform.science.AlgorithmRunner;
+import com.amee.platform.science.CO2Amount;
 import com.amee.platform.science.StartEndDate;
 import com.amee.restlet.AuthorizeResource;
 import com.amee.service.data.DataConstants;
@@ -48,7 +48,6 @@ import org.w3c.dom.Element;
 
 import javax.script.ScriptException;
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -205,7 +204,7 @@ public class AlgorithmResource extends AuthorizeResource implements Serializable
 
         private StringBuffer error = null;
 
-        private BigDecimal amount = null;
+        private Double amount = null;
 
         private String values = null;
 
@@ -230,11 +229,11 @@ public class AlgorithmResource extends AuthorizeResource implements Serializable
             this.error = error;
         }
 
-        public BigDecimal getAmount() {
+        public Double getAmount() {
             return amount;
         }
 
-        public void setAmount(BigDecimal amount) {
+        public void setAmount(Double amount) {
             this.amount = amount;
         }
